+++ ORStools/proc/directions_points_layer_proc.py
@@ -46,6 +46,7 @@ from ..utils.processing import get_params_optimize
 # noinspection PyPep8Naming
 class ORSDirectionsPointsLayerAlgo(ORSBaseProcessingAlgorithm):
     """Algorithm class for Directions Lines."""
+
     def __init__(self):
         super().__init__()
         self.ALGO_NAME = 'directions_from_points_1_layer'
@@ -66,6 +67,8 @@ class ORSDirectionsPointsLayerAlgo(ORSBaseProcessingAlgorithm):
                 name=self.IN_FIELD,
                 description="Layer ID Field",
                 parentLayerParameterName=self.IN_POINTS,
+                defaultValue=None,
+                optional=True
             ),
             QgsProcessingParameterEnum(
                 self.IN_PROFILE,
@@ -103,12 +106,17 @@ class ORSDirectionsPointsLayerAlgo(ORSBaseProcessingAlgorithm):
         )
 
         source_field_name = parameters[self.IN_FIELD]
+        get_fields_options = dict()
+        if source_field_name:
+            get_fields_options.update(
+                from_type=source.fields().field(source_field_name).type(),
+                from_name=source_field_name
+            )
+
+        sink_fields = directions_core.get_fields(**get_fields_options, line=True)
 
         (sink, dest_id) = self.parameterAsSink(parameters, self.OUT, context,
-                                               directions_core.get_fields(
-                                                   from_type=source.fields().field(source_field_name).type(),
-                                                   from_name=source_field_name,
-                                                   line=True),
+                                               sink_fields,
                                                QgsWkbTypes.LineString,
                                                QgsCoordinateReferenceSystem.fromEpsgId(4326))
         count = source.featureCount()
@@ -130,7 +138,7 @@ class ORSDirectionsPointsLayerAlgo(ORSBaseProcessingAlgorithm):
                 for point in feat.geometry().asMultiPoint():
                     points.append(x_former.transform(QgsPointXY(point)))
                 input_points.append(points)
-                from_values.append(feat[source_field_name])
+                from_values.append(feat[source_field_name] if source_field_name else None)
 
         for num, (points, from_value) in enumerate(zip(input_points, from_values)):
             # Stop the algorithm if cancel button has been clicked