+++ ORStools/proc/matrix_proc.py
@@ -66,6 +66,8 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
                 name=self.IN_START_FIELD,
                 description="Start ID Field (can be used for joining)",
                 parentLayerParameterName=self.IN_START,
+                defaultValue=None,
+                optional=True,
             ),
             QgsProcessingParameterFeatureSource(
                 name=self.IN_END,
@@ -76,6 +78,8 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
                 name=self.IN_END_FIELD,
                 description="End ID Field (can be used for joining)",
                 parentLayerParameterName=self.IN_END,
+                defaultValue=None,
+                optional=True,
             ),
             QgsProcessingParameterEnum(
                 self.IN_PROFILE,
@@ -97,7 +101,9 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
             self.IN_START,
             context
         )
+
         source_field_name = parameters[self.IN_START_FIELD]
+        source_field = source.fields().field(source_field_name) if source_field_name else None
 
         destination = self.parameterAsSource(
             parameters,
@@ -105,11 +111,7 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
             context
         )
         destination_field_name = parameters[self.IN_END_FIELD]
-
-        # Get fields from field name
-        source_field = source.fields().field(source_field_name)
-
-        destination_field = destination.fields().field(destination_field_name)
+        destination_field = destination.fields().field(destination_field_name) if destination_field_name else None
 
         # Abort when MultiPoint type
         if (source.wkbType() or destination.wkbType()) == 4:
@@ -152,6 +154,15 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
             'id': 'Matrix'
         }
 
+        # get types of set ID fields
+        field_types = dict()
+        if source_field:
+            field_types.update({"source_type": source_field.type()})
+        if destination_field:
+            field_types.update({"destination_type": destination_field.type()})
+
+        sink_fields = self.get_fields(**field_types)
+
         # Make request and catch ApiError
         try:
             response = ors_client.request('/v2/matrix/' + profile, {}, post_json=params)
@@ -167,15 +178,14 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
             parameters,
             self.OUT,
             context,
-            self.get_fields(
-                source_field.type(),
-                destination_field.type()
-            ),
+            sink_fields,
             QgsWkbTypes.NoGeometry
         )
 
-        sources_attributes = [feat.attribute(source_field_name) for feat in sources_features]
-        destinations_attributes = [feat.attribute(destination_field_name) for feat in destination_features]
+        sources_attributes = [feat.attribute(source_field_name) if source_field_name else feat.id() for feat in
+                              sources_features]
+        destinations_attributes = [feat.attribute(destination_field_name) if destination_field_name else feat.id() for
+                                   feat in destination_features]
 
         for s, source in enumerate(sources_attributes):
             for d, destination in enumerate(destinations_attributes):
@@ -193,8 +203,10 @@ class ORSMatrixAlgo(ORSBaseProcessingAlgorithm):
 
         return {self.OUT: dest_id}
 
+    # TODO working source_type and destination_type differ in both name and type from get_fields in directions_core.
+    #  Change to be consistent
     @staticmethod
-    def get_fields(source_type, destination_type):
+    def get_fields(source_type=QVariant.Int, destination_type=QVariant.Int):
 
         fields = QgsFields()
         fields.append(QgsField("FROM_ID", source_type))